Generator Iterator Type Traits (#17)
I have a personal game engine project that I am currently migrating to use library away from the less robust precursor. For this, the `generator_iterator<T>` needs to satisfy a few properties; namely that is must satisfy `std::is_trivially_destructible` and `std::is_copy_assignable`.

The reason for this is that I have embedded Lua and in order to allow entity iteration in scripts, the iterator gets created directly in the Lua stack. The copy assignable is so that I can assign to the memory allocated in the stack, and the trivially destructible is so that the Lua garbage collector can delete it later.

It may be that these are restrictive requirements, however I don't think that's the case with the current implementation; currently, the `generator_iterator<T>` isn't really an iterator in the classic sense, as the "position" that it is currently at is really stored inside the `generator<T>`. Thus, the iterator is more of a view that can fetch the value and advance, but it is not possible to have two iterators at different points. So copying these things is fine so long as the user knows that all iterators point to the same value, and trivial destruction is fine since it only has a handle back to the generator.

- Add `static_assert`s into the `generator_iterator<T>` constructor to properly codify these requirements.
- `generator_iterator<T>` now stores a pointer to the `generator<T>` rather than a reference since it needs to be reassignable.